add a yices-dynamic flag
Yices comes in two flavors: with gmp statically
or dinamically linked.
The dynamic version should be preferred, in order
to be able to use yices from ghci.
Therefore this flag builds for the dynamically linked
yices by default.

However in non Unix systems it can be necessary to
manually disable this flag